acr-agents/src/agents/RadiologyAgent.py
"""
Radiology Agent for ACR Platform
Specializes in colposcopy image analysis
"""

from typing import Dict, Any, List
import torch
import torch.nn.functional as F
from PIL import Image
import numpy as np
from loguru import logger

from ..base_agent import BaseACRAgent
from ..skills.image_analysis import ImageAnalysisSkill
from ..skills.swrl_reasoner import SWRLReasonerSkill


class RadiologyAgent(BaseACRAgent):
    """
    Radiology Agent for analyzing colposcopy images
    Detects cervical lesions and classifies grades
    """

    def __init__(
        self,
        name: str = "radiology_agent",
        ontology_path: str = "./ontology/ACR.owl",
        model_path: str = None,
        redis_url: str = "redis://localhost:6379",
        port: int = 8001,
    ):
        super().__init__(
            name=name,
            agent_type="RADIOLOGY",
            ontology_path=ontology_path,
            redis_url=redis_url,
            port=port,
        )

        # Initialize skills
        self.image_analysis_skill = ImageAnalysisSkill(model_path)
        self.swrl_reasoner_skill = SWRLReasonerSkill(self.ontology)

        # Load SWRL rules for radiology
        self.swrl_rules = [
            "diagnostic.swrl:Rule8",  # High confidence finding
            "diagnostic.swrl:Rule9",  # Multi-agent agreement
            "diagnostic.swrl:Rule12",  # Inconsistent findings
        ]

        logger.info(f"RadiologyAgent initialized with {len(self.swrl_rules)} SWRL rules")

    def get_capabilities(self) -> List[Dict[str, Any]]:
        """Return radiology agent capabilities"""
        return [
            {
                "id": "colposcopy_analysis",
                "name": "Colposcopy Image Analysis",
                "requiredModalities": ["COLPOSCOPY"],
                "outputTypes": ["CervicalGrade", "RiskLevel"],
            },
            {
                "id": "region_detection",
                "name": "Region of Interest Detection",
                "requiredModalities": ["COLPOSCOPY"],
                "outputTypes": ["BoundingBox", "SegmentationMask"],
            },
        ]

    async def analyze_case(self, case_data: Dict[str, Any]) -> Dict[str, Any]:
        """
        Analyze colposcopy images for a case
        Returns findings with grade, confidence, and regions
        """
        try:
            images = case_data.get("images", [])
            colposcopy_images = [
                img for img in images if img.get("modality_type") == "COLPOSCOPY"
            ]

            if not colposcopy_images:
                logger.warning("No colposcopy images found in case")
                return {
                    "grade": "UNKNOWN",
                    "risk_level": "UNKNOWN",
                    "confidence": 0.0,
                    "features": [],
                    "regions": [],
                }

            # Process the primary colposcopy image
            primary_image = colposcopy_images[0]
            image_path = primary_image.get("local_path")

            logger.info(f"Analyzing colposcopy image: {image_path}")

            # 1. Run image analysis skill
            analysis_result = await self.image_analysis_skill.analyze_image(image_path)

            # 2. Detect regions of interest
            regions = await self.image_analysis_skill.detect_regions(image_path)

            # 3. Extract visual features
            features = self._extract_features(analysis_result)

            # 4. Apply SWRL reasoning
            swrl_result = self.swrl_reasoner_skill.apply_rules(
                {
                    "grade": analysis_result["predicted_grade"],
                    "confidence": analysis_result["confidence"],
                    "features": features,
                }
            )

            # 5. Compile findings
            findings = {
                "grade": analysis_result["predicted_grade"],
                "risk_level": self._map_grade_to_risk(analysis_result["predicted_grade"]),
                "confidence": analysis_result["confidence"],
                "features": features,
                "regions": regions,
                "reasoning": swrl_result.get("reasoning", ""),
                "applied_rules": swrl_result.get("applied_rules", []),
            }

            logger.info(
                f"Radiology analysis complete: Grade={findings['grade']}, Confidence={findings['confidence']:.3f}"
            )

            return findings

        except Exception as e:
            logger.error(f"Error in radiology analysis: {e}")
            raise

    def _extract_features(self, analysis_result: Dict[str, Any]) -> List[str]:
        """Extract visual features from analysis result"""
        features = []

        # Acetowhite changes
        if analysis_result.get("acetowhite_detected", False):
            features.append("acetowhite_epithelium")

        # Vascular patterns
        if analysis_result.get("abnormal_vessels", False):
            features.append("abnormal_vascular_pattern")

        # Lesion characteristics
        if analysis_result.get("lesion_size", 0) > 0.5:
            features.append("large_lesion")

        # Color changes
        if analysis_result.get("iodine_negative", False):
            features.append("iodine_negative_area")

        return features

    def _map_grade_to_risk(self, grade: str) -> str:
        """Map cervical grade to risk level"""
        grade_risk_map = {
            "NORMAL": "LOW",
            "ASCUS": "MODERATE",
            "LSIL": "MODERATE",
            "HSIL": "HIGH",
            "CANCER": "VERY_HIGH",
        }
        return grade_risk_map.get(grade, "UNKNOWN")


async def main():
    """Main entry point for Radiology Agent"""
    agent = RadiologyAgent(
        name="radiology_agent_1",
        ontology_path="./ontology/ACR.owl",
    )

    await agent.run()


if __name__ == "__main__":
    import asyncio

    asyncio.run(main())
