acr-agents/requirements.txt
# ACR Agents - Python Dependencies
# Version: 0.8.0

# Fetch.ai Agent Framework
fetchai==0.13.0
uagents==0.12.0
cosmpy==0.9.2

# Ontology and Reasoning
owlready2==0.46
rdflib==7.0.0

# Machine Learning
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.1.0

# Image Processing
opencv-python==4.8.1.78
scikit-image==0.22.0

# Data Handling
pandas==2.1.3
pydicom==2.4.3

# API Communication
httpx==0.25.2
aiohttp==3.9.1
websockets==12.0

# Redis for message queue
redis==5.0.1
hiredis==2.2.3

# Configuration
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
